backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
reportlab
python-docx
python-multipart
pdfminer.six
pydantic
